Struct openbrush_lang::storage::MultiMapping

pub struct MultiMapping<K, V, TGK = RefGuard<K>, TGV = ValueGuard<V>, KeyType: StorageKey = AutoKey> { /* private fields */ }

A mapping of one key to many values. The mapping provides iteration functionality over all key’s values.

Implementations

impl<K, V, TGK, TGV, KeyType> MultiMapping<K, V, TGK, TGV, KeyType>where K: Packed, V: Packed, KeyType: StorageKey,

pub fn insert<'b>( &'b mut self, key: <TGK as TypeGuard<'b>>::Type, value: &<TGV as TypeGuard<'b>>::Type )where for<'a> TGK: TypeGuard<'a>, for<'a> TGV: TypeGuard<'a>, for<'a> <TGK as TypeGuard<'a>>::Type: Encode + Copy, for<'a> <TGV as TypeGuard<'a>>::Type: Packed,

Insert the given value to the contract storage at key.

pub fn count<'b>(&'b self, key: <TGK as TypeGuard<'b>>::Type) -> u128where for<'a> TGK: TypeGuard<'a>, for<'a> <TGK as TypeGuard<'a>>::Type: Encode,

Returns the count of values stored under the key.

pub fn get_value<'b>( &self, key: <TGK as TypeGuard<'b>>::Type, index: &u128 ) -> Option<V>where for<'a> TGK: TypeGuard<'a>, for<'a> TGV: TypeGuard<'a>, for<'a> <TGK as TypeGuard<'a>>::Type: Encode + Copy,

Get the value at (key, index) from the contract storage.

Returns None if no value exists at the given (key, index).

pub fn get_index<'b>( &self, key: <TGK as TypeGuard<'b>>::Type, value: &<TGV as TypeGuard<'b>>::Type ) -> Option<u128>where for<'a> TGK: TypeGuard<'a>, for<'a> TGV: TypeGuard<'a>, for<'a> <TGK as TypeGuard<'a>>::Type: Encode + Copy, for<'a> <TGV as TypeGuard<'a>>::Type: Packed,

Get the index of (key, value) from the contract storage.

Returns None if no value exists for the given key.

pub fn size_value<'b>( &'b self, key: <TGK as TypeGuard<'b>>::Type, value: &<TGV as TypeGuard<'b>>::Type ) -> Option<u32>where for<'a> TGK: TypeGuard<'a>, for<'a> TGV: TypeGuard<'a>, for<'a> <TGK as TypeGuard<'a>>::Type: Encode + Copy, for<'a> <TGV as TypeGuard<'a>>::Type: Packed,

Get the size of a value stored at (key, value) in the contract storage.

Returns None if no value exists at the given (key, value).

pub fn size_index<'b>( &'b self, key: <TGK as TypeGuard<'b>>::Type, index: &u128 ) -> Option<u32>where for<'a> TGK: TypeGuard<'a>, for<'a> TGV: TypeGuard<'a>, for<'a> <TGK as TypeGuard<'a>>::Type: Encode, for<'a> <TGV as TypeGuard<'a>>::Type: Packed,

Get the size of a value stored at (key, index) in the contract storage.

Returns None if no value exists at the given (key, index).

pub fn contains<'b>(&'b self, key: <TGK as TypeGuard<'b>>::Type) -> boolwhere for<'a> TGK: TypeGuard<'a>, for<'a> <TGK as TypeGuard<'a>>::Type: Encode,

Checks if any value is stored at the given key in the contract storage.

pub fn contains_value<'b>( &'b self, key: <TGK as TypeGuard<'b>>::Type, value: &<TGV as TypeGuard<'b>>::Type ) -> boolwhere for<'a> TGK: TypeGuard<'a>, for<'a> TGV: TypeGuard<'a>, for<'a> <TGK as TypeGuard<'a>>::Type: Encode, for<'a> <TGV as TypeGuard<'a>>::Type: Packed,

Checks if the value is stored at the given key in the contract storage.

pub fn contains_index<'b>( &'b self, key: <TGK as TypeGuard<'b>>::Type, index: &u128 ) -> boolwhere for<'a> TGK: TypeGuard<'a>, for<'a> TGV: TypeGuard<'a>, for<'a> <TGK as TypeGuard<'a>>::Type: Encode, for<'a> <TGV as TypeGuard<'a>>::Type: Packed,

Checks if any value is stored at the given (key, index) in the contract storage.

pub fn remove_value<'b>( &'b mut self, key: <TGK as TypeGuard<'b>>::Type, value: &<TGV as TypeGuard<'b>>::Type )where for<'a> TGK: TypeGuard<'a>, for<'a> TGV: TypeGuard<'a>, for<'a> <TGK as TypeGuard<'a>>::Type: Encode + Copy, for<'a> <TGV as TypeGuard<'a>>::Type: Packed + From<V>,

Clears the value at key from storage.

pub fn remove_index<'b>( &'b mut self, key: <TGK as TypeGuard<'b>>::Type, index: &u128 )where for<'a> TGK: TypeGuard<'a>, for<'a> TGV: TypeGuard<'a>, for<'a> <TGK as TypeGuard<'a>>::Type: Encode + Copy, for<'a> <TGV as TypeGuard<'a>>::Type: Packed + From<V>,

Clears the value at (key, index) from storage.