Module openbrush_lang::traits

Constants

ZERO_ADDRESS

Traits

AccountIdExt

The trait provides some useful methods for AccountId type.

DefaultEnv

Each object has access to default environment via Self::env(). It can be used for interaction with host functions of the blockchain.

Flush

This trait is automatically implemented for storage structs.

OccupiedStorage

Trait describes that the storage KEY already is occupied by WithData type. Implementation of that trait for each storage field prevents the user from occupying the same storage cells.

OccupyStorage

Each upgradeable storage type should occupy its storage key. The trait helps to describe what storage key is occupied by the type.

Storage

Implementation of the trait means that the type stores some Data inside. It is stored in one exemplar, and reference can be retrieved from the object by get or get_mut methods. The trait is helpful for generics implementations when you don’t know precisely the final type, but it is enough for you to know that it has some Data inside.

StorageAsMut

Helper trait for Storage to provide user-friendly API to retrieve data as mutable reference.

StorageAsRef

Helper trait for Storage to provide user-friendly API to retrieve data as reference.

Type Definitions

AccountId

Aliases for types of the default environment

Balance

BlockNumber

ChainExtension

EnvAccess

Hash

String

Timestamp

Derive Macros

Storage

The macro implements openbrush::traits::Storage and openbrush::traits::OccupiedStorage traits for each field marked by #[storage_field] attribute. Each field’s type should implement the openbrush::traits::OccupyStorage trait with a unique storage key. Each occupied storage key should be unique for each type otherwise compilation will fail.